requirements.txt
streamlit>=1.27.0
langchain>=0.0.300
groq>=0.20
google-generative-ai>=0.3.0
python-dotenv>=1.0.0

pytube>=15.0.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2